requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]
python-jose
pydantic
python-multipart
pytest
pytest-asyncio
httpx
